Time: 35 ms (51.47%) | Memory: 61.3 MB (77.94%) - LeetSync

+++ 3351-maximize-happiness-of-selected-children/maximize-happiness-of-selected-children.java
@@ -6,7 +6,7 @@ class Solution {
         for (int i = happiness.length - 1; i >= happiness.length - k; i--) {
 
             if (happiness[i] - count > 0) {
-                ans += (long) (happiness[i] - count);
+                ans +=  happiness[i] - count;
             }
 
             count++;
